(PC-22487)[PRO] feat: new form filter adage
