+++ pro/src/pages/AdageIframe/app/components/OffersInstantSearch/OffersInstantSearch.tsx
@@ -3,6 +3,7 @@ import React, { useContext } from 'react'
 import { Configure, InstantSearch } from 'react-instantsearch-dom'
 
 import { VenueResponse } from 'apiClient/adage'
+import useActiveFeature from 'hooks/useActiveFeature'
 import {
   ALGOLIA_API_KEY,
   ALGOLIA_APP_ID,
@@ -13,6 +14,7 @@ import { FacetFiltersContext } from '../../providers'
 import { AnalyticsContextProvider } from '../../providers/AnalyticsContextProvider'
 
 import { OffersSearch } from './OffersSearch/OffersSearch'
+import { OldOffersSearch } from './OffersSearch/OldOffersSearch'
 
 const searchClient = algoliasearch(ALGOLIA_APP_ID, ALGOLIA_API_KEY)
 const attributesToRetrieve = [
@@ -35,6 +37,8 @@ export const OffersInstantSearch = ({
 }): JSX.Element => {
   const { facetFilters } = useContext(FacetFiltersContext)
 
+  const newAdageFilters = useActiveFeature('WIP_ENABLE_NEW_ADAGE_FILTERS')
+
   return (
     <InstantSearch
       indexName={ALGOLIA_COLLECTIVE_OFFERS_INDEX}
@@ -48,10 +52,20 @@ export const OffersInstantSearch = ({
         hitsPerPage={8}
       />
       <AnalyticsContextProvider>
-        <OffersSearch
-          removeVenueFilter={removeVenueFilter}
-          venueFilter={venueFilter}
-        />
+        {
+          /* istanbul ignore next: DEBT to fix: delete condition with ff */
+          newAdageFilters ? (
+            <OffersSearch
+              removeVenueFilter={removeVenueFilter}
+              venueFilter={venueFilter}
+            />
+          ) : (
+            <OldOffersSearch
+              removeVenueFilter={removeVenueFilter}
+              venueFilter={venueFilter}
+            />
+          )
+        }
       </AnalyticsContextProvider>
     </InstantSearch>
   )
