+++ pro/src/pages/AdageIframe/app/components/OffersInstantSearch/OffersSearch/OfferFilters/OfferFilters.module.scss
@@ -1,0 +1,15 @@
+@use "styles/variables/_colorsAdageIframe.scss" as colors;
+@use "styles/mixins/_rem.scss" as rem;
+@import "styles";
+
+.filter-container {
+  position: relative;
+
+  &-search {
+    position: absolute;
+    right: 0;
+    top: rem.torem(8px);
+    border-bottom-left-radius: 0;
+    border-top-left-radius: 0;
+  }
+}
